recursion

+++ 0x08-recursion/3-factorial.c
@@ -5,7 +5,7 @@
  *
 * Return: factorial number
 */
-int factorial(int n)
+int factorial(long long n)
 {
 if (n == 1)
 {
